requirements.txt
streamlit==1.38.0
langchain==0.3.1
langchain-openai==0.2.1
langchain-core==0.3.7
langchain-community==0.2.17
python-dotenv==1.0.1
pypdf2==3.0.1
python-docx==1.1.2
mermaid-py==0.6.0
cairosvg==2.7.2
pytesseract==0.3.13
pdf2image==1.17.0
pillow==10.4.0